+++ inst/tinytest/test_coordinates.R
@@ -6,15 +6,13 @@ coda <- list(
 )
 
 ## Ternary coordinates
-tern <- isopleuros:::coordinates_ternary(coda)
+tern <- coordinates_ternary(coda)
 expect_equal_to_reference(tern, file = "_snaps/coord_ternary.rds")
 
 ## Cartesian coordinates
-cart <- isopleuros:::coordinates_cartesian(tern)
+cart <- coordinates_cartesian(tern)
 expect_equal_to_reference(cart, file = "_snaps/coord_cartesian.rds")
 expect_equal(coda, cart)
 
 coda$x <- -coda$x
 expect_error(coordinates_ternary(coda))
-tern$x <- -tern$x
-expect_error(coordinates_cartesian(tern))